fix(bug-on-res): changed route, edited model controller and services
