+++ routes/indexRoute.js
@@ -2,7 +2,6 @@ const express = require('express')
 const router = express.Router()
 const userRoute = require('./userRoute')
 const testmiddleware = require('../middlewares/test.middleware')
-const loginRoute = require('./loginRoute')
 const customerRoutes = require("./customerRoutes");
 const transactionRoutes = require("./transactionRoutes");
 const loanRoutes = require("./loanRoute");
@@ -10,8 +9,7 @@ const loanRoutes = require("./loanRoute");
 
 
 
-router.use('/register', userRoute)
-router.use('/login', loginRoute)
+router.use('/users', userRoute)
 router.use("/customers", customerRoutes);
 router.use("/transactions", transactionRoutes);
 router.use("/loans", loanRoutes);
